--- a/benchmarks/benchmark_import_time.py
+++ b/benchmarks/benchmark_import_time.py
@@ -1,0 +1,16 @@
+from subprocess import run, PIPE
+from sys import executable
+
+class ImportSuite:
+    """Benchmark the time it takes to import various modules"""
+    def setup(self):
+        pass
+
+    def time_import_numpy(self):
+        results = run(executable + ' -c "import numpy"',
+            stdout=PIPE, stderr=PIPE, stdin=PIPE, shell=True)
+
+    def time_import_skimage(self):
+        results = run(executable + ' -c "import skimage"',
+            stdout=PIPE, stderr=PIPE, stdin=PIPE, shell=True)
+