--- a/benchmarks/benchmark_import_time.py
+++ b/benchmarks/benchmark_import_time.py
@@ -3,14 +3,18 @@ from sys import executable
 
 class ImportSuite:
     """Benchmark the time it takes to import various modules"""
-    def setup(self):
+    params = [
+        'numpy',
+        'skimage',
+        'skimage.feature',
+        'skimage.morphology',
+        'skimage.color',
+    ]
+    param_names = ["package_name"]
+    def setup(self, package_name):
         pass
 
-    def time_import_numpy(self):
-        results = run(executable + ' -c "import numpy"',
-            stdout=PIPE, stderr=PIPE, stdin=PIPE, shell=True)
-
-    def time_import_skimage(self):
-        results = run(executable + ' -c "import skimage"',
+    def time_import(self, package_name):
+        results = run(executable + ' -c "import ' + package_name + '"',
             stdout=PIPE, stderr=PIPE, stdin=PIPE, shell=True)