--- a/skimage/data/_binary_blobs.py
+++ b/skimage/data/_binary_blobs.py
@@ -1,5 +1,4 @@
 import numpy as np
-from ..filters import gaussian
 
 
 def binary_blobs(length=512, blob_size_fraction=0.1, n_dim=2,
@@ -43,6 +42,10 @@ def binary_blobs(length=512, blob_size_fraction=0.1, n_dim=2,
     >>> # Blobs cover a smaller volume fraction of the image
     >>> blobs = data.binary_blobs(length=256, volume_fraction=0.3)
     """
+    # filters is quite an expensive import since it imports all of scipy.signal
+    # We lazy import here
+    from ..filters import gaussian
+
     rs = np.random.RandomState(seed)
     shape = tuple([length] * n_dim)
     mask = np.zeros(shape)