--- a/skimage/util/_montage.py
+++ b/skimage/util/_montage.py
@@ -1,5 +1,4 @@
 import numpy as np
-from .. import exposure
 
 
 __all__ = ['montage']
@@ -85,6 +84,10 @@ def montage(arr_in, fill='mean', rescale_intensity=False, grid_shape=None,
     (2, 6)
     """
 
+    # exposure imports scipy.linalg which is quite expensive.
+    # Since skimage.util is in the critical import path, we lazy import
+    # exposure to improve import time
+    from .. import exposure
     if multichannel:
         arr_in = np.asarray(arr_in)
     else: