--- a/skimage/util/apply_parallel.py
+++ b/skimage/util/apply_parallel.py
@@ -1,6 +1,3 @@
-from math import ceil
-from multiprocessing import cpu_count
-
 __all__ = ['apply_parallel']
 
 
@@ -21,6 +18,10 @@ def _get_chunks(shape, ncpu):
     >>> _get_chunks((2, 4), 2)
     ((1, 1), (4,))
     """
+    # since apply_parallel is in the critical import path, we lazy import
+    # math just when we need it.
+    from math import ceil
+
     chunks = []
     nchunks_per_dim = int(ceil(ncpu ** (1./len(shape))))
 
@@ -119,6 +120,9 @@ def apply_parallel(function, array, chunks=None, depth=0, mode=None,
     if chunks is None:
         shape = array.shape
         try:
+            # since apply_parallel is in the critical import path, we lazy
+            # import multiprocessing just when we need it.
+            from multiprocessing import cpu_count
             ncpu = cpu_count()
         except NotImplementedError:
             ncpu = 4