## Description
This delays almost all imports from the critical import path of scikit image.

Now, on a "good" install, only numpy is imported, all other libraries are now lazy imported and only used when the user needs them in their code.

This decreases the total import time from 0.7 s to 0.13 s. Most of this is due to importing numpy.
Before:
![image](https://user-images.githubusercontent.com/90008/61596678-a3bd0100-abd4-11e9-9d6e-0777114e2ba3.png)

After:
![image](https://user-images.githubusercontent.com/90008/61596672-97d13f00-abd4-11e9-9ec5-470ec90a11b7.png)


Ref: #4038

## Checklist

<!-- It's fine to submit PRs which are a work in progress! -->
<!-- But before they are merged, all PRs should provide: -->
- [ ] [Docstrings for all functions](https://github.com/numpy/numpy/blob/master/doc/example.py)
- [ ] Gallery example in `./doc/examples` (new features only)
- [ ] Benchmark in `./benchmarks`, if your changes aren't covered by an
  existing benchmark
- [ ] Unit tests
- [ ] Clean style in [the spirit of PEP8](https://www.python.org/dev/peps/pep-0008/)

<!-- For detailed information on these and other aspects see -->
<!-- the scikit-image contribution guidelines. -->
<!-- https://scikit-image.org/docs/dev/contribute.html -->

## For reviewers

<!-- Don't remove the checklist below. -->
- [ ] Check that the PR title is short, concise, and will make sense 1 year
  later.
- [ ] Check that new functions are imported in corresponding `__init__.py`.
- [ ] Check that new features, API changes, and deprecations are mentioned in
      `doc/release/release_dev.rst`.
- [ ] Consider backporting the PR with `@meeseeksdev backport to v0.14.x`