@hmaarrfk 2019-07-22
Thats fine, and I'll gladly make the PR to get it back to this level again. We do all the effort to keep the namespace clean, but we import much of scipy to create optional datasets for examples. It really does take too long to import scikit-image and I am basically having to drop it as one of my dependencies due to that.

I can work on adding a benchmark if that helps.