@pep8speaks 2019-07-22
Hello @hmaarrfk! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:













* In the file [`skimage/util/arraycrop.py`](https://github.com/scikit-image/scikit-image/blob/8ce76909288e9a9fdc89dd9d9425758f130e3ed3/skimage/util/arraycrop.py):

> [Line 42:80](https://github.com/scikit-image/scikit-image/blob/8ce76909288e9a9fdc89dd9d9425758f130e3ed3/skimage/util/arraycrop.py#L42): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (81 > 79 characters)