@jni 2019-07-22
Thanks @hmaarrfk! I support this work, and think we should definitely not be so lackadaisical about this sort of stuff. See also #3397.

With regards to @stefanv's point that you will be forever maintaining this, perhaps adding a benchmark for imports would be a good idea?

My main concern is about how unpythonic this all is (PEP8 puts all imports at the top, which I like), and it makes me wonder whether we can do better by fiddling with `skimage/__init__.py` instead.