@jni 2019-07-22
(but note that I still might be in favour of merging this as-is, just need to think about it a bit more!)