@hmaarrfk 2019-07-22
i totally understand the desire not to have random imports everywhere. Unfortunately, certain things like `platform` take a non-negligeable amount of time to import.

I think you had an issue where we talked about the craziness in the dependencies.

I'm trying to write a benchmark for this, but it isn't as easy as I thought ;)