@hmaarrfk 2019-07-22
some of the other ones are: importing dask array for a convenience tutorial function.

That is fine, but that increases the import time of scikit-image by about 200 ms, which is non negligeable. This will happen when people download distributions such as anaconda, and find `scikit-image` as a result of combined packages being installed at once.

In terms of cost to benefit analysis, I think this is well worth it at this point to keep the package light, especially for the `img_as_*` functions, which are pretty core to any analysis.

Here are some results:

master
```
[100.00%] ··· ============== ===========
               package_name
              -------------- -----------
                  numpy       126±0.9ms
                 skimage       498±2ms
              ============== ===========
```

this branch:

```
[ 75.00%] ··· ============== =========
               package_name
              -------------- ---------
                  numpy       126±1ms
                 skimage      128±1ms
              ============== =========
```

```

       before           after         ratio
     [cba1798f]       [3e5f1a08]
     <master>         <improve_import_time~1>
-        579±90ms         389±30ms     0.67  benchmark_import_time.ImportSuite.time_import('skimage.feature') [mark-xps/conda-py3.7-cython-numpy1.16-scipy]
-         597±8ms         363±20ms     0.61  benchmark_import_time.ImportSuite.time_import('skimage.feature') [mark-xps/conda-py3.7-cython-numpy1.15-scipy]
-        564±20ms         326±20ms     0.58  benchmark_import_time.ImportSuite.time_import('skimage.morphology') [mark-xps/conda-py3.7-cython-numpy1.15-scipy]
-        586±70ms         324±10ms     0.55  benchmark_import_time.ImportSuite.time_import('skimage.morphology') [mark-xps/conda-py3.7-cython-numpy1.16-scipy]
-        509±20ms          162±7ms     0.32  benchmark_import_time.ImportSuite.time_import('skimage.color') [mark-xps/conda-py3.7-cython-numpy1.15-scipy]
-        580±40ms          166±3ms     0.29  benchmark_import_time.ImportSuite.time_import('skimage.color') [mark-xps/conda-py3.7-cython-numpy1.16-scipy]
-         570±4ms          131±5ms     0.23  benchmark_import_time.ImportSuite.time_import('skimage') [mark-xps/conda-py3.7-cython-numpy1.16-scipy]
-         606±5ms          132±2ms     0.22  benchmark_import_time.ImportSuite.time_import('skimage') [mark-xps/conda-py3.7-cython-numpy1.15-scipy]

```