@hmaarrfk 2019-07-22
Numpy seems to have their own benchmarks, though they don't display quite nicely.
https://pv.github.io/numpy-bench/#bench_import.Import.time_fft

Finally, i don't think their numbers are quite accurate, if it really does take 1 second to start up numpy, that is REALLY slow.