@hmaarrfk 2019-07-22
@jni, i wish I knew how to use `Popen`, then we could try `preloading` certain parts of the python standard library and running only the relevant import statement in the benchmark. but as is now, the benchmark shows that we are basically limited by importing numpy, which is fine since this is a `numpy` first library.