@hmaarrfk 2019-07-22
https://snarky.ca/lazy-importing-in-python-3-7/