@stefanv 2019-07-22
I didn't realize we were up to 0.5s already—that is quite painful.  Pity about the style, but I don't see an easy way around it, other than messing with hacks like lazy import (https://pypi.org/project/lazy-import/).