@stefanv 2019-07-22
And, yes, this is a great usage of benchmarks!